move express init outside controllers
